+++ src/pmedian.jl
@@ -1,8 +1,31 @@
 module PMedian
 
 export pmedian
+export pmedian_with_distances
 
 using JuMP, HiGHS
+
+"""
+    PMedianResult
+
+# Fields
+
+- `centers::Vector`: Indices of selected centers.
+- `model::JuMP.Model`: JuMP model.
+- `objective::Float64`: Objective value.
+- `z::Matrix`: Binary matrix of assignments. If z[i, j] = 1 then ith point is connected to the jth point.
+- `y::Vector`: Binary vector. If y[i] is 1 then ith point is a depot.
+"""
+struct PMedianResult
+    centers::Vector
+    model::JuMP.Model
+    objective::Float64
+    z::Matrix
+    y::Vector
+end
+
+
+
 
 function euclidean(u, v)::Float64
     d = (u .- v)
@@ -10,20 +33,25 @@ function euclidean(u, v)::Float64
 end
 
 
+
+
 """
 
     pmedian(data, ncenters)
 
 # Arguments 
+
 - `data::Matrix`: Coordinates of locations 
 - `ncenters::Int`: Number of centers 
 
 # Descriptions 
+
 The function calculates Euclidean distances between all possible rows of the matrix data. 
 `ncenters` locations are then selected that minimizes the total distances to the nearest rows. 
 
 # Output 
-- `Dict{String, Any}`: The dictionary object that holds the results. 
+
+- `PMedianResult`: PMedianResult object. 
 
 # Example 
 
@@ -36,17 +64,20 @@ julia> data3 = rand(10, 2) .+ 100;
 
 julia> data = vcat(data1, data2, data3);
 
-julia> result = pmedian(data, 3)
-Dict{String, Any} with 5 entries:
-  "centers"   => [6, 13, 25]
-  "model"     => A JuMP Model…
-  "objective" => 10.5921
-  "z"         => [-0.0 -0.0 … 0.0 0.0; -0.0 -0.0 … 0.0 0.0; … ; 0.0 0.0 … -0.0 0.0; 0.0 0.0 … 0.0 -0.0]
-  "y"         => [0.0, -0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, -0.0, 0.0  …  0.0, 0.0, 0.0, -0.0, 1.0, 0.0, -0.0,…
+julia> result = pmedian(data, 3);
+
+julia> result.centers
+3-element Vector{Int64}:
+  1
+ 16
+ 21
+
+ julia> result.objective
+ 11.531012240599605
 ``` 
 
 """
-function pmedian(data::Matrix, ncenters::Int)
+function pmedian(data::Matrix, ncenters::Int)::PMedianResult
 
     n, p = size(data)
 
@@ -77,9 +108,9 @@ end
 `ncenters` locations are selected that minimizes the total distances to the nearest rows. 
 
 # Output 
-- `Dict{String, Any}`: The dictionary object that holds the results. 
+- `PMedianResult`: PMedianResult object. 
 """
-function pmedian_with_distances(distancematrix::Matrix, ncenters::Int)
+function pmedian_with_distances(distancematrix::Matrix, ncenters::Int)::PMedianResult
 
     n, _ = size(distancematrix)
 
@@ -108,12 +139,12 @@ function pmedian_with_distances(distancematrix::Matrix, ncenters::Int)
 
     optimize!(model)
 
-    return Dict(
-        "z" => value.(z),
-        "y" => value.(y),
-        "centers" => findall(x -> x == 1, value.(y)),
-        "objective" => JuMP.objective_value(model),
-        "model" => model,
+    return PMedianResult(
+        findall(x -> x == 1, value.(y)),    
+        model,
+        JuMP.objective_value(model),
+        value.(z),
+        value.(y),
     )
 end